requirements.txt
google-api-python-client==1.8.2
numpy==1.18.3
Pillow==7.1.2
scipy==1.4.1
yattag==1.13.2

